Back to stack
I decided to go back to stack. I tested and it appears that 10 and 11
and 12 are faster via recursion, but 13 and 14 are much faster via
stack. It could be an issue with where the goal lays in the order. Its
on the heap, so some performance is lost, but I think this works well.